fix bug

+++ DESCrack.cpp
@@ -45,7 +45,7 @@ int main(int argc,char*argv[])
 	{
 		keyNum = argc - 3;
 		for(index = 0;index < keyNum;index++)
-			p_cs -> AddKey(atoi(argv[index+3]));
+			p_cs -> AddKey(atoll(argv[index+3]));
 	}
 	else
 	{
