+++ Verified.cpp
@@ -12,8 +12,8 @@ void Usage()
 int main(int argc,char*argv[])
 {
 	FILE * file;
-	long long chainLen, chainCount, fileLen;
-	RainbowChain chain; long long index;
+	uint64_t chainLen, chainCount, fileLen;
+	RainbowChain chain; uint64_t index;
 	ChainWalkContext cwc;
 
 	if(argc != 3)
@@ -30,7 +30,7 @@ int main(int argc,char*argv[])
 
 	fseek(file, 0, SEEK_SET);
 
-	chainLen = atoi(argv[2]);
+	chainLen = atoll(argv[2]);
 	fileLen  = GetFileLen(file);
 	
 	if(fileLen % 16 != 0)
